fix Version fields type to uint64.

+++ models/read_transaction.go
@@ -2,6 +2,7 @@ package models
 
 import (
 	"encoding/json"
+
 	"github.com/tidwall/gjson"
 )
 
@@ -82,7 +83,7 @@ type SuiTransactionBlock struct {
 
 type SuiObjectRef struct {
 	ObjectId string `json:"objectId"`
-	Version  int    `json:"version"`
+	Version  uint64 `json:"version"`
 	Digest   string `json:"digest"`
 }
 
@@ -97,7 +98,7 @@ type SuiGasData struct {
 type SuiObjectChangePublished struct {
 	Type      string   `json:"type"`
 	PackageId string   `json:"packageId"`
-	Version   int      `json:"version"`
+	Version   uint64   `json:"version"`
 	Digest    string   `json:"digest"`
 	Modules   []string `json:"modules"`
 }
@@ -108,7 +109,7 @@ type SuiObjectChangeTransferred struct {
 	Recipient  ObjectOwner `json:"recipient"`
 	ObjectType string      `json:"objectType"`
 	ObjectId   string      `json:"objectId"`
-	Version    int         `json:"version"`
+	Version    uint64      `json:"version"`
 	Digest     string      `json:"digest"`
 }
 
@@ -118,8 +119,8 @@ type SuiObjectChangeMutated struct {
 	Owner           ObjectOwner `json:"owner"`
 	ObjectType      string      `json:"objectType"`
 	ObjectId        string      `json:"objectId"`
-	Version         int         `json:"version"`
-	PreviousVersion int         `json:"previousVersion"`
+	Version         uint64      `json:"version"`
+	PreviousVersion uint64      `json:"previousVersion"`
 	Digest          string      `json:"digest"`
 }
 
@@ -128,7 +129,7 @@ type SuiObjectChangeDeleted struct {
 	Sender     string `json:"sender"`
 	ObjectType string `json:"objectType"`
 	ObjectId   string `json:"objectId"`
-	Version    int    `json:"version"`
+	Version    uint64 `json:"version"`
 }
 
 type SuiObjectChangeWrapped struct {
@@ -136,7 +137,7 @@ type SuiObjectChangeWrapped struct {
 	Sender     string `json:"sender"`
 	ObjectType string `json:"objectType"`
 	ObjectId   string `json:"objectId"`
-	Version    int    `json:"version"`
+	Version    uint64 `json:"version"`
 }
 
 type SuiObjectChangeCreated struct {
@@ -145,7 +146,7 @@ type SuiObjectChangeCreated struct {
 	Owner      ObjectOwner `json:"owner"`
 	ObjectType string      `json:"objectType"`
 	ObjectId   string      `json:"objectId"`
-	Version    int         `json:"version"`
+	Version    uint64      `json:"version"`
 	Digest     string      `json:"digest"`
 }
